requirements.txt
zipencrypt
zipfile
argparse
pycryptodome